requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
pycodestyle==2.12.1
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.2.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
passlib~=1.7.4
python-jose~=3.4.0
loguru~=0.7.3
requests~=2.32.3
passlib[bcrypt]
bcrypt>=4.0.0